ALMSTK-33: Cast file_info.st_size, rather than trying to print it directly
The type for 'st_size' is defined as __off_t and changes depending on the
target we are building. To ensure we can always compile simply cast the
value to a uint32_t and print it accordingly.

+++ cmsg/src/cmsg.c
@@ -836,8 +836,8 @@ cmsg_get_msg_from_file (const ProtobufCMessageDescriptor *desc, const char *file
     if (len != file_info.st_size)
     {
         CMSG_LOG_GEN_ERROR
-            ("Failed to read packed message data (expected = %llu, read = %zu).",
-             file_info.st_size, len);
+            ("Failed to read packed message data (expected = %u, read = %zu).",
+             (uint32_t) file_info.st_size, len);
         free (packed_data);
         fclose (fp);
         return NULL;
